refactor(client/notes): re-organize all note files to be contained in one Notes feature directory
