+++ client/src/features/Notes/index.js
@@ -1,0 +1,10 @@
+// Export components
+export * from "./components";
+
+// Export pages
+export * from "./pages";
+
+// Export services
+export { default as noteService } from "./services/noteService";
+
+// Export hooks
