+++ client/src/pages/Dashboard.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { DashboardHeader } from "@/components/DashboardHeader";
 import { Sidebar } from "@/components/Sidebar";
 import { MiniScreenTaskBoard } from "@/features/Tasks/pages/MiniScreenTaskBoard";
-import { NotesPanel } from "@/components/NotesPanel";
+import { NotesPanel } from "@/features/Notes/components/NotesPanel";
 import PomodoroContainer from "@/features/Pomodoro";
 import { ThemeToggle } from "@/components/ThemeToggle";
